+++ src/accml/core/model/result.py
@@ -43,3 +43,29 @@ class Result:
     data: Sequence[ReadTogether]
     #: in the expected view / context as produced by the backend
     orig_data: Sequence[ReadTogether]
+
+
+
+def single_reading_deserializer(obj: dict, cls, **kwargs):
+    assert cls == SingleReading, "only prepared to handle single reading"
+    name = obj["name"]
+    payload = obj["payload"]
+    return cls(name=name, payload=payload)
+
+
+def dataclass_delegate_to_jsons_method(obj, **kwargs):
+    import jsons
+    r =  jsons.dump(obj.to_jsons(), **kwargs)
+    return r
+
+
+def register_deserializers_to_json_fork(t_fork):
+    import jsons
+    jsons.set_deserializer(single_reading_deserializer, SingleReading, high_prio=True, fork_inst=t_fork)
+
+
+def register_serializers_to_json_fork(t_fork):
+    import jsons
+    jsons.set_serializer(dataclass_delegate_to_jsons_method, ReadTogether, high_prio=True, fork_inst=t_fork)
+
+
